+++ src/client/java/rh/maparthelper/compat/ModMenuIntegration.java
@@ -1,10 +1,11 @@
-package rh.maparthelper.config;
+package rh.maparthelper.compat;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import me.shedaniel.autoconfig.AutoConfig;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import rh.maparthelper.config.MapartHelperConfig;
 
 @Environment(EnvType.CLIENT)
 public class ModMenuIntegration implements ModMenuApi {