+++ src/main/java/rh/maparthelper/config/palette/PaletteColors.java
@@ -6,11 +6,14 @@ import rh.maparthelper.colors.ColorUtils;
 import rh.maparthelper.colors.MapColorEntry;
 
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 
 public class PaletteColors {
     static final Map<Integer, MapColorEntry> argbMapColors = new HashMap<>();
     static final Map<Integer, MapColorEntry> cachedClosestColors = new HashMap<>();
+    static final Set<MapColor> excludingColors = new HashSet<>();
 
     public static MapColorEntry getMapColorEntryByARGB(int argb) {
         if (argb == 0) return MapColorEntry.CLEAR;
@@ -28,6 +31,7 @@ public class PaletteColors {
             rgbOriginal = ColorUtils.getRGB(argb);
 
         for (MapColor color : PaletteConfigManager.presetsConfig.getCurrentPresetColors()) {
+            if (excludingColors.contains(color)) continue;
             for (int brightId = 0; brightId < 3; brightId++) {
                 MapColor.Brightness brightness;
                 brightness = color == MapColor.WATER_BLUE ? MapColor.Brightness.NORMAL : MapColor.Brightness.validateAndGet(brightId);
@@ -69,6 +73,7 @@ public class PaletteColors {
             rgbOriginal = ColorUtils.getRGB(argb);
 
         for (MapColor color : PaletteConfigManager.presetsConfig.getCurrentPresetColors()) {
+            if (excludingColors.contains(color)) continue;
             int current = color.getRenderColor(MapColor.Brightness.NORMAL);
             if (current == argb) return new MapColorEntry(color, MapColor.Brightness.NORMAL, new int[]{0, 0, 0});
 
@@ -100,7 +105,19 @@ public class PaletteColors {
         return cachedClosestColors.computeIfAbsent(argb, c -> PaletteColors.getClosestColor2D(c, useDithering));
     }
 
-    public static void clearColorCache(){
+    public static int excludingColorsAmount() {
+        return excludingColors.size();
+    }
+
+    public static boolean addExcludingColors(Set<MapColor> excludingColors) {
+        return PaletteColors.excludingColors.addAll(excludingColors);
+    }
+
+    public static void clearExcludingColors() {
+        excludingColors.clear();
+    }
+
+    public static void clearColorCache() {
         cachedClosestColors.clear();
         ColorUtils.clearRgb2LabCache();
     }